+++ cypress/e2e/models/administration/jira-connection/jira.ts
